app/requirements.txt
Flask==1.1.1
requests==2.22.0
RPi.GPIO==0.7.0
